Added clean param to build master index
- A clean aparam was added to prevent the deletion of the index when importing
  packets

+++ app/pql/pcapfile.py
@@ -1,3 +1,4 @@
+import logging
 import sqlite3
 from struct import pack, unpack
 
@@ -5,9 +6,6 @@ import pql.packet_index as pkt_index
 from config.config import Config
 from packet.layers.packet_decode import PacketDecode
 from packet.layers.packet_hdr import PktHeader
-
-import logging
-
 
 log = logging.getLogger("packetdb")
 
@@ -173,28 +171,32 @@ class PcapFile:
 
         conn.close()
 
-    def build_master_index(self, master_index):
+    def build_master_index(self, master_index, clean=False):
         db_name = f"{Config.pcap_master_index()}"
         conn = sqlite3.connect(db_name)
         c = conn.cursor()
-        c.execute("drop table if exists master_index;")
-        c.execute("""
-                  create table if not exists master_index (
-                      id integer primary key autoincrement,
-                      start_ts integer not null,
-                      end_ts integer not null,
-                      file_id integer not null
-                      );
-                  """)
+
+        if clean:
+            c.execute("drop table if exists master_index;")
+
+            c.execute("""
+                      create table if not exists master_index (
+                          id integer primary key autoincrement,
+                          start_ts integer not null,
+                          end_ts integer not null,
+                          file_id integer not null
+                          );
+                      """)
         c.execute('''PRAGMA synchronous = EXTRA''')
         c.execute('''PRAGMA journal_mode = WAL''')
         c.executemany(
             "INSERT INTO master_index (start_ts, end_ts, file_id) VALUES (?,?,?)", master_index)
         conn.commit()
 
-        c.execute("""
-                  create index idx_timestamp
-                  on master_index (start_ts, end_ts);
-                  """)
+        if clean:
+            c.execute("""
+                      create index if not exists  idx_timestamp
+                      on master_index (start_ts, end_ts);
+                      """)
 
         conn.close()
